move frame_interface into own header
