+++ src/frame.h
@@ -3,62 +3,15 @@
 
 #pragma once
 
+#include "core/frame-interface.h"
 #include "core/frame-continuation.h"
 #include "core/frame-additional-data.h"
-#include "core/extension.h"
 #include "basics.h"
 #include <atomic>
-#include <array>
 #include <vector>
-#include <map>
-#include <math.h>
+//#include <math.h>
 
 namespace librealsense {
-
-
-class stream_profile_interface;
-class archive_interface;
-class sensor_interface;
-
-
-class frame_interface
-{
-public:
-    virtual frame_header const & get_header() const = 0;
-    virtual rs2_metadata_type get_frame_metadata( const rs2_frame_metadata_value & frame_metadata ) const = 0;
-    virtual bool supports_frame_metadata( const rs2_frame_metadata_value & frame_metadata ) const = 0;
-    virtual int get_frame_data_size() const = 0;
-    virtual const uint8_t * get_frame_data() const = 0;
-    virtual rs2_time_t get_frame_timestamp() const = 0;
-    virtual rs2_timestamp_domain get_frame_timestamp_domain() const = 0;
-    virtual void set_timestamp( double new_ts ) = 0;
-    virtual unsigned long long get_frame_number() const = 0;
-
-    virtual void set_timestamp_domain( rs2_timestamp_domain timestamp_domain ) = 0;
-    virtual rs2_time_t get_frame_system_time() const = 0;
-    virtual std::shared_ptr< stream_profile_interface > get_stream() const = 0;
-    virtual void set_stream( std::shared_ptr< stream_profile_interface > sp ) = 0;
-
-    virtual std::shared_ptr< sensor_interface > get_sensor() const = 0;
-    virtual void set_sensor( std::shared_ptr< sensor_interface > ) = 0;
-
-    virtual void acquire() = 0;
-    virtual void release() = 0;
-    virtual frame_interface * publish( std::shared_ptr< archive_interface > new_owner ) = 0;
-    virtual void unpublish() = 0;
-    virtual void attach_continuation( frame_continuation && continuation ) = 0;
-    virtual void disable_continuation() = 0;
-    virtual archive_interface * get_owner() const = 0;
-
-    virtual void mark_fixed() = 0;
-    virtual bool is_fixed() const = 0;
-    virtual void set_blocking( bool state ) = 0;
-    virtual bool is_blocking() const = 0;
-
-    virtual void keep() = 0;
-
-    virtual ~frame_interface() = default;
-};
 
 
 struct LRS_EXTENSION_API frame_holder
